refactor: change "goinfre" denomination to be "volatile"
This makes it more easily uderstundable for everyone.
It used to be called Goinfre because that's the name of the partition the apps were saved to locally (`~/goinfre`).
Even for a technical standpoint it's better to save apps to `/tmp`, so might as well rename the vars to volatile for ease of understanding.